refactor: SearchInput 파일 이동 및 리렌더링 개선 (#20)
